move ISimpleVector3
