+++ src/app/shared/models/render/base.render-model.ts
@@ -2,12 +2,8 @@ import { Group, Vector3 } from "three";
 import { BaseMaterialService } from "../../services/3d-managers/base-material.service";
 import { FileModelType } from "../model-type.enum";
 import { Observable } from "rxjs";
-
-export type ISimpleVector3 = {
-  readonly x: number,
-  readonly y: number,
-  readonly z: number,
-}
+import { ISimpleVector3 } from "../simple-types";
+import { BaseModelManifest } from "../model-manifest";
 
 export abstract class BaseRenderModel<T extends FileModelType> {
