+++ src/app/shared/models/simple-types.ts
@@ -1,0 +1,5 @@
+export type ISimpleVector3 = {
+  readonly x: number,
+  readonly y: number,
+  readonly z: number,
+}
